Moved display related functions in a separate file.
This fix an issue where the lcd is initialized twice during setup. A singleton is now used for both the servo and the main file. This change also encapsulate the graphic libraries in the display python file to avoid polluing the main namespace with them for nothing.
The logic to check for the logo in jpg format was removed since it will be always the case with git update.
The sleep that was part of some display functions has been moved out back to the main files because this is not display specific.